:recycle: Flyttet scrollfade ut i egen komponent
